Create a "flask proxy" that call the final service in body["destination_url"]. Then returns the answer moving the data to body["body"] and headers to body["headers"]. Thats why is impossible read headers like x-subject-token where the token id is returned.
But now, I cannot parse json answers!
* In Second Life: http://wiki.secondlife.com/wiki/Json_usage_in_LSL
* In OpenSim there is a different (and better) implementation to parse json files: http://opensimulator.org/wiki/JsonStore_Module
So writing a rest client in lsl is not the same in OpenSim and Second Life.

Next step: use OpenSim/Second Life as little as possible to interact with external services.
I will implement in the flask app all logic and use the OpenSim/Second Life Object to store results returned as simple csv strings.

Architecture:
* FrontEnd: OpenSim/Second Life stores Object information.
* FackEnd (FrontEnd of the Backend): Flask App receives calls from FrontEnd, modify and resend to OpenStack. Then returns a simple string easy to parse from OpenSim/Second Life lsl.
* Backend: Receives FackEnd calls and does the final tasks.